+++ MOP/src/Common/ConsoleCommands.cs
@@ -106,40 +106,9 @@ namespace MOP.Common
                         return;
                     }
 
-                    if (RulesManager.Instance.IgnoreRules.Count > 0)
+                    foreach (Rule rule in RulesManager.Instance.GetList())
                     {
-                        ModConsole.Log("<color=yellow><b>Ignore Rules</b></color>");
-                        foreach (IgnoreRule r in RulesManager.Instance.IgnoreRules)
-                            ModConsole.Log($"<b>Object:</b> {r.ObjectName}" + (r.TotalIgnore ? " <i>(Full Ignore)</i>" : "") + $" <b>By</b> {r.Mod?.ID}");
-                    }
-
-                    if (RulesManager.Instance.IgnoreRulesAtPlaces.Count > 0)
-                    {
-                        ModConsole.Log("\n<color=yellow><b>Ignore Rules At Place</b></color>");
-                        foreach (IgnoreRuleAtPlace r in RulesManager.Instance.IgnoreRulesAtPlaces)
-                            ModConsole.Log($"<b>Place:</b> {r.Place} <b>Object:</b> {r.ObjectName}" + $" <b>By</b> {r.Mod?.ID}");
-                    }
-
-                    if (RulesManager.Instance.ToggleRules.Count > 0)
-                    {
-                        ModConsole.Log("\n<color=yellow><b>Toggle Rules</b></color>");
-                        foreach (ToggleRule r in RulesManager.Instance.ToggleRules)
-                            ModConsole.Log($"<b>Object:</b> {r.ObjectName} <b>Toggle Mode:</b> {r.ToggleMode}" + $" <b>By</b> {r.Mod?.ID}");
-                    }
-
-                    if (RulesManager.Instance.NewSectors.Count > 0)
-                    {
-                        ModConsole.Log("\n<color=yellow><b>New Sectors</b></color>");
-                        foreach (NewSector r in RulesManager.Instance.NewSectors)
-                            ModConsole.Log($"<b>Pos:</b> {r.Position} <b>Scale:</b> {r.Scale} <b>Rot:</b> {r.Rotation} " +
-                                $"<b>Ignore:</b> {string.Join(", ", r.Whitelist)}" + $" <b>By</b> {r.Mod?.ID}");
-                    }
-
-                    if (RulesManager.Instance.ChangeParentRules.Count > 0)
-                    {
-                        ModConsole.Log("\n<color=yellow><b>Change Parent</b></color>");
-                        foreach (ChangeParentRule r in RulesManager.Instance.ChangeParentRules)
-                            ModConsole.Log($"<b>Object:</b> {r.ObjectName} <b>New Parent:</b> {r.NewParentName}" + $" <b>By</b> {r.Mod?.ID}");
+                        ModConsole.Log($"<b>{rule.GetType().Name}</b> {rule}");
                     }
 
                     ModConsole.Log("\n<color=yellow><b>Special Rules</b></color>");
@@ -307,8 +276,8 @@ namespace MOP.Common
                     if (args.Length > 1)
                     {
                         if (RulesManager.Instance.LoadRules &&
-                            (RulesManager.Instance.IgnoreRules.Count > 0 || RulesManager.Instance.IgnoreRulesAtPlaces.Count > 0 ||
-                             RulesManager.Instance.NewSectors.Count > 0 || RulesManager.Instance.ToggleRules.Count > 0))
+                            (RulesManager.Instance.GetList<IgnoreRule>().Count > 0 || RulesManager.Instance.GetList<IgnoreRuleAtPlace>().Count > 0 ||
+                             RulesManager.Instance.GetList<NewSector>().Count > 0 || RulesManager.Instance.GetList<ToggleRule>().Count > 0))
                         {
                             ModConsole.Log("<color=red>WARNING:</color> For accurate results, use \"mop load-rules false\" to prevent MOP from using rule files.");
                         }