+++ MOP/src/Rules/Types/ToggleRule.cs
@@ -29,5 +29,10 @@ namespace MOP.Rules.Types
             this.ObjectName = ObjectName;
             this.ToggleMode = ToggleMode;
         }
+
+        public override string ToString()
+        {
+            return base.ToString() + $" Object: {ObjectName} Mode: {ToggleMode}";
+        }
     }
 }