+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 
 - Improved MOP_Log generation
 - MOP will now detect, if engine was assembled using third-party tools (I mean, it's really weird when the save file says that the engine block is assembled, and it's bolted, but all bolts have a value of 0)
-- Optimization to Debug Monitor
+- Optimization of Debug Monitor
+- Optimization of MOP rules
 
 ### Bug Fixes